Start the process to replace mock sunburst data with our data
