+++ .babelrc
@@ -1,14 +1,15 @@
 {
-  presets: [
+  "presets": [
     [
-      '@babel/preset-env',
+      "@babel/preset-env",
       {
-        modules: false
+        "modules": false
       }
     ],
-    '@babel/preset-react'
+    "@babel/preset-react"
   ],
-  plugins: [
-    'react-hot-loader/babel'
+  "plugins": [
+    "react-hot-loader/babel",
+    "@babel/plugin-transform-runtime"
   ]
 }
