+++ package.json
@@ -18,6 +18,7 @@
   },
   "dependencies": {
     "@nivo/sunburst": "^0.67.0",
+    "axios": "^0.21.1",
     "bcrypt": "^4.0.1",
     "body-parser": "^1.19.0",
     "bootstrap": "^4.6.0",
@@ -42,6 +43,7 @@
   },
   "devDependencies": {
     "@babel/core": "^7.13.10",
+    "@babel/plugin-transform-runtime": "^7.13.15",
     "@babel/preset-env": "^7.13.12",
     "@babel/preset-react": "^7.12.13",
     "@hot-loader/react-dom": "^17.0.1+4.13.0",
