+++ src/components/categoriesNTagsSunBurst/index.js
@@ -1,0 +1,1 @@
+export {CategoriesNTagsSunBurst} from "./CategoriesNTagsSunBurst";
