+++ webpack.config.js
@@ -49,7 +49,6 @@ const config = {
     alias: {
       // react: path.resolve('./node_modules/react'),
       "react-dom": "@hot-loader/react-dom",
-      
     },
   },
   plugins: [
